MDL-26113 Lesson: temporary fix of page content alignment
That rendering method should not use cell style attribute at all.
Instead, the alignment should be controlled by CSS. However, this will
require more massive changes in the Lesson code to be done by the module
maintainer. As a temporary fix, left alignment is set as it usually
matches the layout that students see.