MDL-26914 Added new language pack installer class and unit tests for it
The new class wraps some common coding patterns to install a language
pack via component_installer. Most notably it deals with the parental
dependency of the language packs.

The new class should also be the only place holding the information
about the location of language packs for the given Moodle version.