MDL-20636 My go and creating a renderer for mod/quiz/review.php.
Note that this breaks the attempt.php page; there are a lot of methods that are missing PHPdoc comment; there are probably now some unused methods that could be deleted; and there are a couple of places which could be handled a bit better.

But, overall, I think this way makes sense.