MDL-28022 glossary entry without attachment can be exported to Mahara as Leap2A content
The problem here was that the manifest file was written into the
'./files/leap2a.xml' instead of the './leap2a.xml' in the ZIP package.
I believe there was a bug in the write_new_file() using the 3rd parameter
incorrectly. A detailed review of all other modules using the same
method must happen soon.