MDL-29794 Shared grading form templates can be deleted now
There is a new capability to manage all shared templates. Without this
capability, the user is allowed to delete just templates they previously
shared.

In the future, an option to edit the template directly might be added.
At the moment, the workaround is to pick the template into a temporary
assignment, edit it there and re-save it as a new template.