MDL-29794 Do not offer re-sharing of unmodified forms
The patch introduces a new field 'timecopied' that holds the timestamp
of when the given form was cloned last time. Using this field, we can
decide whether the form can be actually re-shared or not.

Note that the logic just hides the icon. It is meant as a usability
feature, not any real protection from sharing.