MDL-29920 User report may display the grade analysis link, too
The original idea was to use the link in the row header (the activity
link). However, for both unified user experience and the overall patch
design reasons, I finally decided not to alter the activity link but to
append the analysis icon. The user experience should be the same as in
the grader report.