MDL-29624 Media embedding system, part 3: resource module (and URL, lesson)
This change makes the resource module use the new API when it embeds media
files (audio/video/Flash), and removes the resourcelib library functions
which were previously used for this.

The old API is still used for embedding non-media files such as PDFs and
HTML.

The URL and Lesson module both used copy/paste duplicates of a section of
resource module code which called the removed functions. I have changed
these too. As a bonus this significantly reduces the duplication involved.