MDL-34118 question backup: a mod can use usages in different ways.
An activity module might want to use question_usages in different ways.
We should suppport this, because it lets people make interesting
activities.

At the moment, the backup and restore code does not support this,
because it uses fixed element names for the question_usage element and
its children. This fix changes things so that a prefix can be appended
to all the element names.