MDL-25012 - blogs: introduce a specific 'enabled' config setting
Rather than overloading the $CFG->bloglevel setting which had a
confusing UI in the appearance subsystem.

In order to achieve this we modify take the defaults from the existing
bloglevel setting and set that for $CFG->enableblogs. Note that in order
to prevent a bad default settings from being set we also set
$CFG->bloglevel to a valid 'enabled' setting.