MDL-34657 datalib: function to generate user sort sql.
The purpose of this method is to ensure that when we display lists of
users in different places, the default sort order is consistent.

Having this code centralised in one place them makes it possible to
improve things accross the board, and there is an example of that here.
There is logic so that if we are searching for particular string, then
users with an exact match are sorted first.