MDL-37770 Assignment submission comments do not work without javascript.
This is mostly fixed by MDL-36832 which correctly sets the $PAGE->url.

This patch hides the submit and textareas from submission comments in the grading table when
javascript is disabled and quickgrading is enabled. This particular combination results in a form
nested in a form which does not work (and cannot be fixed in a reasonable amount of time).