MDL-38507 JavaScript: Move definition of YUI_config before setup of SimpleYUI
Since moving to SimpleYUI, the order of inclusion for the YUI_config
variable is more important.

Simple YUI defines a global instance of YUI so that it can be used without
specifying the full YUI().use() syntax. As a result, any configuration
applied when that global instance is setup must exist already.