MDL-39552 Make behat code re-entrant safe
With yui combo in action, some times setup.php is
included more than once. So it's mandatory to
guarantee that code until ABORT_AFTER_CONFIG_CANCEL
can be executed safely multiple times in the same
request.

This change just prevents the behat section to be executed
more than once. All the values that are set within it
should be already defined for the re-entrant call.