MDL-40103 stats: Move db temp templates to inline xmldb tables
Move all statistics temporary tables into the code.  We are not
going to continue to use database templates.  They were a mix
of reading the install.xml and tables in the database.  It is
much clearer to have them created inline at the point they are
used in the code.