MDL-39851 javascript: Full screen dialogs on small screens
This change adds support to M.core.dialogue for showing fullscreen
dialogs on small screen sizes. There are 2 new configuration values that
can be set to control this when creating a dialog (responsive and responsiveWidth).

Setting responsive to false will disable this functionality completely (old behaviour).
Setting responsiveWidth controls the screen width that the dialog will switch to
fullscreen mode. Try not to think of the device that will be accessing the dialog,
but the width at which the content of the dialog needs to be displayed differently.

The activity chooser and the filepicker are the two dialogs that have been converted
in this patch. The filepicker first had to be converted to use M.core.dialog and not
Y.Panel.