MDL-39851 javascript: Fixes for filepicker dialogs
Fixes display issues with the filepicker dialogs that were
converted to M.core.dialogue.

Note: These dialogue use table based layouts and the table width is not
counted in the form that it sits in. I have changed the dialogues to have
a wider width but for really small screens this causes horizontal scrolling.

The only solution for that is to rewrite all the filepicker/filemanager dialogs
to use css for layout.