MDL-41188 javascript: Change full screen dialogue to position: fixed
This fixes all sorts of bother with the fullscreen dialogues. It also means that
you don't lose your scroll position in the document when the full screen popup
opens/closes.

It prevents several bugs on mobile where the dialogue could get scrolled off
the page (and handles orientation changes better).

Contains one other tiny tweak to make the close button on the dialogue
a bit bigger so it's easier to click (only for fullscreen).