MDL-40045 login: removed add_to_log call when a user updates their password
This add_to_log call is not being replaced by a user_updated event as the
call to $userauth->user_update_password should (and does in the case of
manual authentication - see MDL-41170) do this already when the password
is changed in the user table. If the authentication plugin is changing
the password externally then there is no need to call the user_updated
event as this should only be called whenever any update is made on the
user table.