MDL-42117 JavaScript: Don't rely on node data for course toolboxes
The YUI versions of getData and setData do not store the data in the DOM,
but in an instance of YUI. If you make modifications in one instance, but
apply them in another, then the data values are not accessible between
these two instances.

Additionally, there's no benefit to storing the instances on
initialisation - if anything, it increases the load overhead.

This change moves toolboxes to calculating their ancestors on click in the
same way it was done in the initialiser.