MDL-42711 theme port from LM: vertical alignment of the course icon
When the course name spans over multiple lines, vertically centered icon does
not look nice (or at least our beloved Benevolent Dictator for Life does not
like it ;-). This hack alignes the icon vertically with the first line of the
course name all the time (or at least it tries to).