MDL-34353 Javascript: Allow chooser dialogues to fall back to non-JS version
In order for a chooser dialogue to be able to fall back to it's non-JS
equivalent, we need to use a different set of styles.

This change makes the chooser dialogue preparation function add a new class
to the body which is utilised in the CSS to accomplish this. When the class
is not present, the dialogue content is displayed inline.